node: Move node name to database
This commit creates the framework for moving node specific attribues to
the database. It also moves the node name to the database and used the
node name as primary key to match between the database backed node
object and the chef backed node object. If a new node gets detected it
will also create the database entry for this node. In case the database
node object is not able to handle the reqest it will automatically fall
back to the chef backed node object.
