+++ crowbar_framework/app/models/api/upgrade.rb
@@ -487,7 +487,7 @@ module Api
 
       # Delete existing pacemaker resources, from other node in the cluster
       def delete_pacemaker_resources(node_name)
-        node = ::Node.find_node_by_name node_name
+        node = ::Node.find_by_name(node_name)
         if node.nil?
           raise_upgrade_error(
             "Node #{node_name} was not found, cannot delete pacemaker resources."
