+++ crowbar_framework/app/models/crowbar_service.rb
@@ -55,11 +55,11 @@ class CrowbarService < ServiceObject
     if answer.first != 200
       found_errors = []
       Node.find("state:crowbar_upgrade").each do |node|
-        error = node["crowbar_wall"]["chef_error"] || ""
+        error = node.crowbar_wall["chef_error"] || ""
         next if error.empty?
         found_errors.push error
         @logger.error("Chef run ended with an error on the node #{node.name}: #{error}")
-        node["crowbar_wall"]["chef_error"] = ""
+        node.crowbar_wall["chef_error"] = ""
         node.save
       end
       unless found_errors.empty?
@@ -98,7 +98,7 @@ class CrowbarService < ServiceObject
       step = "wait_for_openstack_shutdown" if node.roles.include? "database-config-default"
 
       # mark the position in the upgrade process
-      node["crowbar_wall"]["crowbar_upgrade_step"] = step
+      node.crowbar_wall["crowbar_upgrade_step"] = step
       node.save
     end
 
@@ -121,7 +121,7 @@ class CrowbarService < ServiceObject
       end
 
       # mark the position in the upgrade process
-      node["crowbar_wall"]["crowbar_upgrade_step"] = step
+      node.crowbar_wall["crowbar_upgrade_step"] = step
       node.save
     end
 
@@ -146,7 +146,7 @@ class CrowbarService < ServiceObject
 
     Node.find("state:crowbar_upgrade AND roles:database-config-default").each do |node|
       # mark the position in the upgrade process
-      node["crowbar_wall"]["crowbar_upgrade_step"] = "db_shutdown"
+      node.crowbar_wall["crowbar_upgrade_step"] = "db_shutdown"
       node.save
     end
 
@@ -156,7 +156,7 @@ class CrowbarService < ServiceObject
     # Finally, set the upgrade step to the point where no further action is done
     # even when the upgrade recipes are accidentally executed
     Node.find("state:crowbar_upgrade AND roles:database-config-default").each do |node|
-      node["crowbar_wall"]["crowbar_upgrade_step"] = "done_openstack_shutdown"
+      node.crowbar_wall["crowbar_upgrade_step"] = "done_openstack_shutdown"
       node.save
     end
 
@@ -184,10 +184,10 @@ class CrowbarService < ServiceObject
     node = nil
 
     with_lock "BA-LOCK" do
-      node = Node.find_node_by_name name
+      node = Node.find_by_name(name)
       if node.nil? and (state == "discovering" or state == "testing")
         @logger.debug("Crowbar transition: creating new node for #{name} to #{state}")
-        node = Node.create_new name
+        node = Node.find_or_create_by(name: name)
         self.transition_save_node = true
       end
       if node.nil?
@@ -222,7 +222,6 @@ class CrowbarService < ServiceObject
         self.transition_save_node = true
         pop_it = true
       end
-
       node.save if transition_save_node
     end
 
@@ -240,7 +239,7 @@ class CrowbarService < ServiceObject
       if Crowbar::Product::is_ses?
         # For SUSE Enterprise Storage, default all non-admin nodes to the right platform
         if state == "discovering" and !node.admin?
-          node["target_platform"] = Crowbar::Product::ses_platform
+          node.target_platform = Crowbar::Product::ses_platform
           node.save
         end
       end
@@ -330,7 +329,7 @@ class CrowbarService < ServiceObject
         # for Hyper-V nodes, only change the state, but do not run chef-client
         node.set_state("crowbar_upgrade")
       else
-        node["crowbar_wall"]["crowbar_upgrade_step"] = "crowbar_upgrade"
+        node.crowbar_wall["crowbar_upgrade_step"] = "crowbar_upgrade"
         node.save
         nodes_to_upgrade.push node.name
       end
@@ -387,8 +386,8 @@ class CrowbarService < ServiceObject
     admin_node = Node.admin_node
 
     upgrade_nodes.each do |node|
-      node["target_platform"] = admin_node["provisioner"]["default_os"]
-      node["crowbar_wall"]["crowbar_upgrade_step"] = "prepare-os-upgrade"
+      node.target_platform = admin_node["provisioner"]["default_os"]
+      node.crowbar_wall["crowbar_upgrade_step"] = "prepare-os-upgrade"
       # skip initial keystone bootstrapping
       if node["run_list_map"].key? "keystone-server"
         node["keystone"]["bootstrap"] = true
@@ -413,7 +412,7 @@ class CrowbarService < ServiceObject
     upgrade_nodes_failed = []
 
     upgrade_nodes.each do |node|
-      node["target_platform"] = admin_node["provisioner"]["default_os"]
+      node.target_platform = admin_node["provisioner"]["default_os"]
       node.save
       node.set_state("os-upgrading")
     end
@@ -423,7 +422,7 @@ class CrowbarService < ServiceObject
     pxecfg_subdir = "bios/pxelinux.cfg"
 
     upgrade_nodes.each do |node|
-      boot_ip_hex = node["crowbar"]["boot_ip_hex"]
+      boot_ip_hex = node.crowbar["crowbar"]["boot_ip_hex"]
       node_arch = node["kernel"]["machine"]
       pxe_conf = "#{discovery_dir}/#{node_arch}/#{pxecfg_subdir}/#{boot_ip_hex}"
       ready_for_reboot = false
@@ -462,7 +461,7 @@ class CrowbarService < ServiceObject
     Node.all.each do |node|
       next unless node.state == "crowbar_upgrade"
       # revert nodes to previous state; mark the wall so apply does not change state again
-      node["crowbar_wall"]["crowbar_upgrade_step"] = "revert_to_ready"
+      node.crowbar_wall["crowbar_upgrade_step"] = "revert_to_ready"
       node.save
       node.set_state("ready")
     end
@@ -478,7 +477,7 @@ class CrowbarService < ServiceObject
   def apply_role_pre_chef_call(old_role, role, all_nodes)
     @logger.debug("crowbar apply_role_pre_chef_call: entering #{all_nodes.inspect}")
     all_nodes.each do |n|
-      node = Node.find_node_by_name n
+      node = Node.find_by_name(n)
       # value of crowbar_wall["crowbar_upgrade"] indicates that the role should be executed
       # but node state should not be changed: this is needed when reverting node state to ready
       if node.role?("crowbar-upgrade") &&
