+++ crowbar_framework/app/models/ntp_service.rb
@@ -55,7 +55,7 @@ class NtpService < ServiceObject
   def transition(inst, name, state)
     @logger.debug("NTP transition: entering: #{name} for #{state}")
 
-    node = Node.find_node_by_name name
+    node = Node.find_by_name(name)
     if node.allocated? && !node.role?("ntp-server")
       db = Proposal.where(barclamp: @bc_name, name: inst).first
       role = RoleObject.find_role_by_name "#{@bc_name}-config-#{inst}"
