+++ crowbar_framework/app/models/provisioner_service.rb
@@ -77,15 +77,15 @@ class ProvisionerService < ServiceObject
 
     if state == "hardware-installing"
       # ensure target platform is set before we claim a disk for boot OS
-      node = Node.find_node_by_name(name)
-      if node[:target_platform].nil? or node[:target_platform].empty?
-        node[:target_platform] = Node.default_platform
+      node = Node.find_by_name(name)
+      if node.target_platform.nil? or node.target_platform.empty?
+        node.target_platform = Node.default_platform
         node.save
       end
     end
 
     if state == "readying"
-      node = Node.find_node_by_name(name)
+      node = Node.find_by_name(name)
       node.process_raid_claims
     end
 
@@ -94,14 +94,14 @@ class ProvisionerService < ServiceObject
       # after a reset.
       @logger.debug("Provisioner transition: clearing node data (claimed disks, boot device, etc.)")
 
-      node = Node.find_node_by_name(name)
+      node = Node.find_by_name(name)
       save_it = false
 
-      node["crowbar_wall"] ||= {}
+      node.crowbar_wall ||= {}
 
       ["boot_device", "claimed_disks"].each do |key|
-        next unless node["crowbar_wall"].key?(key)
-        node["crowbar_wall"].delete(key)
+        next unless node.crowbar_wall.key?(key)
+        node.crowbar_wall.delete(key)
         save_it = true
       end
 
@@ -110,13 +110,13 @@ class ProvisionerService < ServiceObject
 
     if state == "delete"
       # BA LOCK NOT NEEDED HERE.  NODE IS DELETING
-      node = Node.find_node_by_name(name)
+      node = Node.find_by_name(name)
       node.crowbar["state"] = "delete-final"
       node.save
     end
 
     # test state machine and call chef-client if state changes
-    node = Node.find_node_by_name(name)
+    node = Node.find_by_name(name)
     role = RoleObject.find_role_by_name "#{@bc_name}-config-#{inst}"
 
     unless node.admin? ||
