+++ crowbar_framework/db/migrate/20161219151456_create_nodes.rb
@@ -1,0 +1,9 @@
+class CreateNodes < ActiveRecord::Migration
+  def change
+    create_table :nodes do |t|
+      t.string :name
+
+      t.timestamps
+    end
+  end
+end
