+++ crowbar_framework/spec/controllers/crowbar_controller_spec.rb
@@ -22,6 +22,8 @@ describe CrowbarController do
 
   before do
     Proposal.where(barclamp: "crowbar", name: "default").first_or_create(barclamp: "crowbar", name: "default")
+    Node.where(name: "testing.crowbar.com").first_or_create(name: "testing.crowbar.com")
+    Node.where(name: "admin.crowbar.com").first_or_create(name: "admin.crowbar.com")
     allow_any_instance_of(CrowbarService).to receive(:apply_role).and_return([200, "OK"])
   end
 
@@ -66,41 +68,37 @@ describe CrowbarController do
 
   describe "POST transition" do
     it "does not allow invalid states" do
-      post :transition, barclamp: "crowbar", id: "default", state: "foobarz", name: "testing"
+      post :transition, barclamp: "crowbar", id: "default", state: "foobarz", name: "testing.crowbar.com"
       expect(response).to be_bad_request
     end
 
     it "does not allow upcased states" do
-      post :transition, barclamp: "crowbar", id: "default", state: "Discovering", name: "testing"
+      post :transition, barclamp: "crowbar", id: "default", state: "Discovering", name: "testing.crowbar.com"
       expect(response).to be_bad_request
     end
 
     it "transitions the node into desired state" do
       allow_any_instance_of(RoleObject).to receive(:find_roles_by_search).and_return([])
-      post :transition, barclamp: "crowbar", id: "default", state: "discovering", name: "testing"
+      post :transition, barclamp: "crowbar", id: "default", state: "discovering", name: "testing.crowbar.com"
       expect(response).to be_success
     end
 
     it "returns plain text message if transitioning fails" do
       allow_any_instance_of(CrowbarService).to receive(:transition).and_return([500, "error"])
-      post :transition, barclamp: "crowbar", id: "default", state: "discovering", name: "testing"
+      post :transition, barclamp: "crowbar", id: "default", state: "discovering", name: "testing.crowbar.com"
       expect(response).to be_server_error
       expect(response.body).to be == "error"
     end
 
     it "returns node as a hash on success when passed a name" do
-      allow_any_instance_of(CrowbarService).to receive(:transition).
-        and_return([200, { name: "testing" }])
-      post :transition, barclamp: "crowbar", id: "default", state: "discovering", name: "testing"
+      post :transition, barclamp: "crowbar", id: "default", state: "discovering", name: "testing.crowbar.com"
       expect(response).to be_success
       json = JSON.parse(response.body)
       expect(json["name"]).to be == "testing.crowbar.com"
     end
 
     it "returns node as a hash on success when passed a node (backward compatibility)" do
-      allow_any_instance_of(CrowbarService).to receive(:transition).
-        and_return([200, Node.find_node_by_name("testing").to_hash])
-      post :transition, barclamp: "crowbar", id: "default", state: "discovering", name: "testing"
+      post :transition, barclamp: "crowbar", id: "default", state: "discovering", name: "testing.crowbar.com"
       expect(response).to be_success
       json = JSON.parse(response.body)
       expect(json["name"]).to be == "testing.crowbar.com"
