+++ crowbar_framework/spec/controllers/machines_controller_spec.rb
@@ -19,7 +19,8 @@ require "spec_helper"
 
 describe MachinesController do
   before do
-    allow_any_instance_of(Node).to receive(:system).and_return(true)
+    allow_any_instance_of(ChefNode).to receive(:system).and_return(true)
+    Node.where(name: "testing.crowbar.com").first_or_create(name: "testing.crowbar.com")
   end
 
   describe "GET index" do
@@ -68,7 +69,7 @@ describe MachinesController do
 
     context "without nodes" do
       before do
-        allow(Node).to receive(:find_all_nodes).and_return([])
+        allow(Node).to receive(:all).and_return(Node.none)
       end
 
       it "renders json" do
@@ -88,18 +89,18 @@ describe MachinesController do
 
   describe "GET show" do
     it "is successful" do
-      get :show, name: "testing", format: "json"
+      get :show, name: "testing.crowbar.com", format: "json"
       expect(response).to have_http_status(:ok)
     end
 
     it "renders json" do
-      get :show, name: "testing", format: "json"
+      get :show, name: "testing.crowbar.com", format: "json"
       expect(JSON.parse(response.body)).to be_a(Hash)
     end
 
     context "for existent node" do
       it "fetches with name" do
-        get :show, name: "testing", format: "json"
+        get :show, name: "testing.crowbar.com", format: "json"
         json = JSON.parse(response.body)
 
         expect(json["name"]).to eq("testing.crowbar.com")
@@ -126,7 +127,7 @@ describe MachinesController do
       it "assignes role compute" do
         expect_any_instance_of(Node).to receive(:intended_role=).with("compute")
 
-        post :role, name: "testing", role: "compute", format: "json"
+        post :role, name: "testing.crowbar.com", role: "compute", format: "json"
         expect(response).to have_http_status(:ok)
       end
     end
@@ -140,7 +141,7 @@ describe MachinesController do
       allow_any_instance_of(Node).to receive(:save).and_return(false)
       expect_any_instance_of(Node).to receive(:intended_role=).with("compute")
 
-      post :role, name: "testing", role: "compute", format: "json"
+      post :role, name: "testing.crowbar.com", role: "compute", format: "json"
 
       expect(response).to have_http_status(:unprocessable_entity)
     end
@@ -151,7 +152,7 @@ describe MachinesController do
       it "renames a node to tester" do
         expect_any_instance_of(Node).to receive(:alias=).with("tester")
 
-        post :rename, name: "testing", alias: "tester", format: "json"
+        post :rename, name: "testing.crowbar.com", alias: "tester", format: "json"
         expect(response).to have_http_status(:ok)
       end
     end
@@ -165,7 +166,7 @@ describe MachinesController do
       allow_any_instance_of(Node).to receive(:save).and_return(false)
       expect_any_instance_of(Node).to receive(:alias=).with("tester")
 
-      post :rename, name: "testing", alias: "tester", format: "json"
+      post :rename, name: "testing.crowbar.com", alias: "tester", format: "json"
 
       expect(response).to have_http_status(:unprocessable_entity)
     end
@@ -178,9 +179,9 @@ describe MachinesController do
     describe "POST #{action}" do
       context "for existent node" do
         it "invokes #{action}" do
-          expect_any_instance_of(Node).to receive(action)
+          expect_any_instance_of(ChefNode).to receive(action)
 
-          post action, name: "testing", format: "json"
+          post action, name: "testing.crowbar.com", format: "json"
           expect(response).to have_http_status(:ok)
         end
       end
@@ -208,15 +209,15 @@ describe MachinesController do
     describe "POST #{action}" do
       context "for existent node" do
         it "invokes #{action}" do
-          expect_any_instance_of(Node).to receive(action)
+          expect_any_instance_of(ChefNode).to receive(action)
 
-          post action, name: "testing", format: "json"
+          post action, name: "testing.crowbar.com", format: "json"
           expect(response).to have_http_status(:ok)
         end
 
         it "return 403 (forbidden) http status for admin node" do
-          allow_any_instance_of(Node).to receive(:admin?).and_return(true)
-          post action, name: "testing", format: "json"
+          allow_any_instance_of(ChefNode).to receive(:admin?).and_return(true)
+          post action, name: "testing.crowbar.com", format: "json"
           expect(response).to have_http_status(:forbidden)
         end
       end
